+++ src/itemCheckbox/ItemCheckbox.style.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+import { Item } from '../_internals/item'
+
+export const StyledItemCheckbox = styled(Item)`
+  cursor: ${props => (props.disabled ? 'default' : 'pointer')};
+
+  & input {
+    position: absolute;
+    clip: rect(0, 0, 0, 0);
+  }
+`